Experimental, not ready for merging to trunk.
Initial support for putting the final build output outside the SDK directory.
This makes it much easier to not get broken builds due to old output in the SDK
in the face of multiple versions of the same source being build. Adds a global
setting for that output directory and a new parameter to SymbianProgram:
'user_libraries' for libs that should be searched for in the final output
directory (instead of the SDK directory). Puts app registration files in the
import directory as well so that apps work on emulated MMC.

This makes most of the 'constants' in arguments.py referenced through the
module in scons_symbian.py so that they can be changed after importing the
module.

Only does WINSCW for now: need to make the same changes to GCCE.

Comments on whether this is the way this should be done requested.